refactor(protocol): change minCheckpointDelay type from uint48 to uint16
Change minCheckpointDelay from uint48 to uint16 to optimize storage packing.
- Max value of uint16 (65535 seconds) = ~18 hours
- Current value (384 seconds) = 1 epoch
- Sufficient for checkpoint rate limiting purposes

+++ packages/protocol/contracts/layer1/shasta/iface/IInbox.sol
@@ -41,7 +41,7 @@ interface IInbox {
         /// @notice The maximum number of checkpoints to store in ring buffer
         uint16 maxCheckpointHistory;
         /// @notice The minimum delay between checkpoints in seconds
-        uint48 minCheckpointDelay;
+        uint16 minCheckpointDelay;
         /// @notice The multiplier to determine when a forced inclusion is too old so that proposing
         /// becomes permissionless
         uint8 permissionlessInclusionMultiplier;
